requirements.txt
django >= 3.2, < 3.3
djangorestframework >= 3.12, < 3.13
psycopg2-binary>=2.9.9
django-filter>=23.5
